+++ ICD.Connect.Settings/Comparers/OriginatorIdComparer.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using ICD.Common.Utils.Comparers;
+using ICD.Connect.Settings.Originators;
 
 namespace ICD.Connect.Settings.Comparers
 {