+++ ICD.Connect.Settings/Cores/AbstractCore.cs
@@ -1,6 +1,7 @@
-﻿using System;
+using System;
+using ICD.Connect.Settings.Originators;
 
-namespace ICD.Connect.Settings.Core
+namespace ICD.Connect.Settings.Cores
 {
 	public abstract class AbstractCore<TSettings> : AbstractOriginator<TSettings>, ICore
 		where TSettings : ICoreSettings, new()