+++ ICD.Connect.Settings/Cores/ICoreSettings.cs
@@ -1,6 +1,6 @@
 ﻿using ICD.Connect.Settings.Header;
 
-namespace ICD.Connect.Settings.Core
+namespace ICD.Connect.Settings.Cores
 {
 	public interface ICoreSettings : ISettings
 	{