+++ ICD.Connect.Settings/Header/ConfigurationHeader.cs
@@ -7,7 +7,7 @@ namespace ICD.Connect.Settings.Header
 {
 	public sealed class ConfigurationHeader
 	{
-		private static readonly Version s_CurrentConfigVersion = new Version("3.0");
+		private static readonly Version s_CurrentConfigVersion = new Version("3.1");
 
 		private const string CONFIG_VERSION_ELEMENT = "ConfigVersion";
 		private const string GENERATED_ON_ELEMENT = "GeneratedOn";