+++ ICD.Connect.Settings/Originators/Simpl/AbstractSimplOriginator.cs
@@ -1,4 +1,4 @@
-﻿namespace ICD.Connect.Settings.Simpl
+﻿namespace ICD.Connect.Settings.Originators.Simpl
 {
 	public abstract class AbstractSimplOriginator<TSettings> : AbstractOriginator<TSettings>, ISimplOriginator
 		where TSettings : ISimplOriginatorSettings, new()