+++ ICD.Connect.Settings/SPlusShims/AbstractSPlusOriginatorShim.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 using ICD.Common.Properties;
 using ICD.Common.Utils.Extensions;
 using ICD.Common.Utils.Services;
+using ICD.Connect.Settings.Cores;
+using ICD.Connect.Settings.Originators;
+using ICD.Connect.Settings.Originators.Simpl;
 using ICD.Common.Utils.Services.Logging;
 using ICD.Connect.API.Commands;
 using ICD.Connect.API.Nodes;
-using ICD.Connect.Settings.Core;
-using ICD.Connect.Settings.Simpl;
 using ICD.Connect.Settings.SPlusShims.GlobalEvents;
 
 namespace ICD.Connect.Settings.SPlusShims