+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added config migration for port configuration
+ - Added Log method to AbstractOriginator for exceptions
+
+### Changed
+ - Moving originators and settings classes into subdirectories
+ - Incremented config version from 3.0 to 3.1
 
 ## [7.3.0] - 2019-01-02
 ### Added